backend/requirements.txt
certifi==2018.4.16
Django==2.0.6
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==2.6.11
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
wincertstore==0.2
